requirements.txt
requests==2.*
pydantic==1.*
python-dateutil==2.*
shapely==1.*